+++ src/main/java/com/ddudu/application/port/in/repeat_ddudu/CreateRepeatDduduUseCase.java
@@ -1,6 +1,6 @@
 package com.ddudu.application.port.in.repeat_ddudu;
 
-import com.ddudu.application.dto.repeat_ddudu.requset.CreateRepeatDduduRequest;
+import com.ddudu.application.dto.repeat_ddudu.request.CreateRepeatDduduRequest;
 
 public interface CreateRepeatDduduUseCase {
